+++ var/lib/torrentwatch-xa/lib/twxa_test_parser.php
@@ -1,8 +1,12 @@
 <?php
 
 // simple pseudo-unit-tester for the parsing and matching engine
-require_once '/var/lib/torrentwatch-xa/lib/twxa_tools.php';
-require_once '/var/lib/torrentwatch-xa/lib/twxa_parse.php';
+
+error_reporting(E_ERROR | E_WARNING | E_PARSE);
+
+// NOTE: we should include config.php, but it is not necessary for testing parsing
+require_once 'twxa_tools.php';
+require_once 'twxa_parse.php';
 $title = $argv[1];
 $title = preg_replace("/<span .*>(.*)<\/span>/", "$1", $title);
 print "\nParsing exactly what is between >>> and <<< on the next line:\n>>>$title<<<\n\n";