+++ var/www/html/torrentwatch-xa/javascript/configure.js
@@ -1,9 +1,0 @@
-// Hide Search -- By Text
-$("input#hideSearchText").on("keyup", function () {
-    var hideSearchText = $(this).val().toLowerCase();
-    $("#hideListContainer li").addClass('hidden').each(function () {
-        if ($(this).find("span.hiddenItem").text().toLowerCase().match(hideSearchText)) {
-            $(this).removeClass('hidden');
-        }
-    })
-});